Split hic2gcmap into two files
